requirements.txt
flask
requests
python-dotenv
fuzzywuzzy
textblob
spacy
apscheduler